feat: standardise all apis to follow new rules about avoiding unnecessary url parameters and keeping dependant objects in path
reverts commits 6edf27ac & 112c9d84

+++ src/tests/projects/memberships/test_api.py
@@ -479,41 +479,23 @@ async def test_create_project_role_ok(client, project_template):
 
 
 #########################################################################
-# PUT /projects/<project_id>/roles/<role_slug>
+# PUT /projects/roles/<role_id>
 #########################################################################
 
 
-async def test_update_project_role_anonymous_user(client, project_template):
-    project = await f.create_project(project_template)
+async def test_update_project_role_anonymous_user(client):
     data = {"permissions": [choices.ProjectPermissions.VIEW_STORY.value]}
 
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{NOT_EXISTING_B64ID}", json=data
-    )
+    response = await client.put(f"/projects/roles/{NOT_EXISTING_B64ID}", json=data)
 
     assert response.status_code == 401, response.data
 
 
-async def test_update_project_role_project_not_found(client):
-    user = await f.create_user()
-    data = {"permissions": [choices.ProjectPermissions.VIEW_STORY.value]}
-
-    client.login(user)
-    response = await client.put(
-        f"/projects/{NOT_EXISTING_B64ID}/roles/{NOT_EXISTING_B64ID}", json=data
-    )
-
-    assert response.status_code == 404, response.data
-
-
 async def test_update_project_role_role_not_found(client, project_template):
-    project = await f.create_project(project_template)
     data = {"permissions": [choices.ProjectPermissions.VIEW_STORY.value]}
 
-    client.login(project.created_by)
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{NOT_EXISTING_B64ID}", json=data
-    )
+    client.login(await f.create_user())
+    response = await client.put(f"/projects/roles/{NOT_EXISTING_B64ID}", json=data)
 
     assert response.status_code == 404, response.data
 
@@ -525,9 +507,7 @@ async def test_update_project_role_user_forbidden_not_member(client, project_tem
     data = {"permissions": [choices.ProjectPermissions.VIEW_STORY.value]}
 
     client.login(user)
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{role.b64id}", json=data
-    )
+    response = await client.put(f"/projects/roles/{role.b64id}", json=data)
 
     assert response.status_code == 403, response.data
 
@@ -547,9 +527,7 @@ async def test_update_project_role_user_forbidden_no_permission(
     data = {"permissions": [choices.ProjectPermissions.VIEW_STORY.value]}
 
     client.login(user)
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{role.b64id}", json=data
-    )
+    response = await client.put(f"/projects/roles/{role.b64id}", json=data)
     assert response.status_code == 403, response.data
 
 
@@ -559,9 +537,7 @@ async def test_update_project_role_role_non_editable(client, project_template):
     data = {"permissions": [choices.ProjectPermissions.VIEW_STORY.value]}
 
     client.login(project.created_by)
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{role.b64id}", json=data
-    )
+    response = await client.put(f"/projects/roles/{role.b64id}", json=data)
 
     assert response.status_code == 403, response.data
 
@@ -572,9 +548,7 @@ async def test_update_project_role_incompatible_permissions(client, project_temp
     data = {"permissions": [choices.ProjectPermissions.MODIFY_STORY.value]}
 
     client.login(project.created_by)
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{role.b64id}", json=data
-    )
+    response = await client.put(f"/projects/roles/{role.b64id}", json=data)
 
     assert response.status_code == 422, response.data
 
@@ -585,9 +559,7 @@ async def test_update_project_role_not_valid_permissions(client, project_templat
     data = {"permissions": ["not_valid", "foo"]}
 
     client.login(project.created_by)
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{role.b64id}", json=data
-    )
+    response = await client.put(f"/projects/roles/{role.b64id}", json=data)
 
     assert response.status_code == 422, response.data
 
@@ -607,9 +579,7 @@ async def test_update_project_role_ok(client, project_template):
     data = {"permissions": [choices.ProjectPermissions.VIEW_STORY.value]}
 
     client.login(pj_member)
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{role.b64id}", json=data
-    )
+    response = await client.put(f"/projects/roles/{role.b64id}", json=data)
     assert response.status_code == 200, response.data
     res = response.json()
     assert data["permissions"] == res["permissions"]
@@ -622,18 +592,14 @@ async def test_update_project_role_ok(client, project_template):
         ],
         "name": "New member",
     }
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{role.b64id}", json=data
-    )
+    response = await client.put(f"/projects/roles/{role.b64id}", json=data)
     assert response.status_code == 200, response.data
     res = response.json()
     assert data["permissions"] == res["permissions"]
     assert data["name"] == res["name"]
     role = await project.roles.aget(slug=res["slug"])
     data = {"name": "New member 2"}
-    response = await client.put(
-        f"/projects/{project.b64id}/roles/{role.b64id}", json=data
-    )
+    response = await client.put(f"/projects/roles/{role.b64id}", json=data)
     assert response.status_code == 200, response.data
     res = response.json()
     assert len(res["permissions"]) == 2
@@ -641,37 +607,19 @@ async def test_update_project_role_ok(client, project_template):
 
 
 #########################################################################
-# DELETE /projects/<project_id>/roles/<role_slug>
+# DELETE /projects/roles/<role_id>
 #########################################################################
 
 
-async def test_delete_project_role_anonymous_user(client, project_template):
-    project = await f.create_project(project_template)
-    role_slug = "member"
-
-    response = await client.delete(f"/projects/{project.b64id}/roles/{role_slug}")
+async def test_delete_project_role_anonymous_user(client):
+    response = await client.delete(f"/projects/roles/{NOT_EXISTING_B64ID}")
 
     assert response.status_code == 401, response.data
 
 
-async def test_delete_project_role_project_not_found(client):
-    user = await f.create_user()
-
-    client.login(user)
-    response = await client.delete(
-        f"/projects/{NOT_EXISTING_B64ID}/roles/{NOT_EXISTING_B64ID}"
-    )
-
-    assert response.status_code == 404, response.data
-
-
 async def test_delete_project_role_role_not_found(client, project_template):
-    project = await f.create_project(project_template)
-
-    client.login(project.created_by)
-    response = await client.delete(
-        f"/projects/{project.b64id}/roles/{NOT_EXISTING_B64ID}"
-    )
+    client.login(await f.create_user())
+    response = await client.delete(f"/projects/roles/{NOT_EXISTING_B64ID}")
 
     assert response.status_code == 404, response.data
 
@@ -681,7 +629,7 @@ async def test_delete_project_role_user_forbidden_not_member(client, project_tem
     project = await f.create_project(project_template)
     role = await project.roles.aget(slug="member")
     client.login(user)
-    response = await client.delete(f"/projects/{project.b64id}/roles/{role.b64id}")
+    response = await client.delete(f"/projects/roles/{role.b64id}")
 
     assert response.status_code == 403, response.data
 
@@ -700,7 +648,7 @@ async def test_delete_project_role_user_forbidden_no_permission(
     await f.create_project_membership(user=user, project=project, role=member_role)
 
     client.login(user)
-    response = await client.delete(f"/projects/{project.b64id}/roles/{role.b64id}")
+    response = await client.delete(f"/projects/roles/{role.b64id}")
     assert response.status_code == 403, response.data
 
 
@@ -708,7 +656,7 @@ async def test_delete_project_role_role_non_editable(client, project_template):
     project = await f.create_project(project_template)
     role = await project.roles.aget(slug="admin")
     client.login(project.created_by)
-    response = await client.delete(f"/projects/{project.b64id}/roles/{role.b64id}")
+    response = await client.delete(f"/projects/roles/{role.b64id}")
 
     assert response.status_code == 403, response.data
 
@@ -718,7 +666,7 @@ async def test_delete_project_role_not_exists_move_to(client, project_template):
     role = await project.roles.aget(slug="member")
     client.login(project.created_by)
     response = await client.delete(
-        f"/projects/{project.b64id}/roles/{role.b64id}?moveTo={NOT_EXISTING_B64ID}"
+        f"/projects/roles/{role.b64id}?moveTo={NOT_EXISTING_B64ID}"
     )
 
     assert response.status_code == 400, response.data
@@ -729,9 +677,7 @@ async def test_delete_project_role_same_role_move_to(client, project_template):
     role = await project.roles.aget(slug="member")
 
     client.login(project.created_by)
-    response = await client.delete(
-        f"/projects/{project.b64id}/roles/{role.b64id}?moveTo={role.b64id}"
-    )
+    response = await client.delete(f"/projects/roles/{role.b64id}?moveTo={role.b64id}")
 
     assert response.status_code == 400, response.data
 
@@ -750,13 +696,13 @@ async def test_delete_project_role_move_to_owner(client, project_template):
 
     client.login(user)
     response = await client.delete(
-        f"/projects/{project.b64id}/roles/{role.b64id}?moveTo={owner_role.b64id}"
+        f"/projects/roles/{role.b64id}?moveTo={owner_role.b64id}"
     )
     assert response.status_code == 403, response.data
 
     client.login(project.created_by)
     response = await client.delete(
-        f"/projects/{project.b64id}/roles/{role.b64id}?moveTo={owner_role.b64id}"
+        f"/projects/roles/{role.b64id}?moveTo={owner_role.b64id}"
     )
     assert response.status_code == 204, response.data
 
@@ -775,7 +721,7 @@ async def test_delete_project_role_ok(client, project_template):
     role = await project.roles.aget(slug="member")
 
     client.login(pj_member)
-    response = await client.delete(f"/projects/{project.b64id}/roles/{role.b64id}")
+    response = await client.delete(f"/projects/roles/{role.b64id}")
     assert response.status_code == 204, response.data
 
 
@@ -794,7 +740,7 @@ async def test_delete_project_role_ok_move_to(client, project_template):
     role_member = await project.roles.aget(slug="member")
     client.login(pj_member)
     response = await client.delete(
-        f"/projects/{project.b64id}/roles/{role_id}?moveTo={role_member.b64id}"
+        f"/projects/roles/{role_id}?moveTo={role_member.b64id}"
     )
     assert response.status_code == 204, response.data
 
@@ -813,5 +759,5 @@ async def test_delete_project_role_ko_move_to_required(client, project_template)
     role_id = general_admin_role.b64id
 
     client.login(pj_member)
-    response = await client.delete(f"/projects/{project.b64id}/roles/{role_id}")
+    response = await client.delete(f"/projects/roles/{role_id}")
     assert response.status_code == 400, response.data
